exhibits(出品)コントローラー作成
