+++ app/helpers/exhibits_helper.rb
@@ -1,0 +1,2 @@
+module ExhibitsHelper
+end
