+++ spec/requests/exhibits_request_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe "Exhibits", type: :request do
+
+end
